+++ examples/file_upload/README.md
@@ -1,0 +1,14 @@
+# File Upload Example
+
+This example allows the user to select a file from their file system.
+The contents of the selected file are then rendered to the page either as a whole or in chunks.
+
+## Concepts
+
+Demonstrates the use of `ReaderService`.
+
+## Improvements
+
+- Show a progress bar if the file is read in chunks
+- Do something interesting with the uploaded file like displaying pictures
+- Improve the presentation of the example with CSS.
